Restructuring and cleanup.
